+++ trace_decoder/src/processed_block_trace_mpt.rs
